fix: enforced_latency -> Delphi scheme, rm horizon w/ expected error
`enforced_latency` should be according to Delphi `time_value` scheme, so it
should be 5d back (Fri) rather than 4d back (Sat) from the Wednesday forecast
due date.

Remove horizon -1 as it likely may trigger errors in the forecasters, and is no
longer evaluated or visualized. (We may want to submit a degenerate distribution
in its place reflecting what we think the latest value is, for purposes of
FluSight runners checking "calibration". But that is not handled in this
commit.)