pull freshest data from healthdata.gov
Use data from healthdata.gov, rather than from the COVIDcast API, so that
we can use the Wednesday update in our training data. The healthdata.gov
data needs a bit of extra processing -- we need to convert to 7dav, and
generate national data (making sure that missing regions, e.g. territories,
are replaced in the sum with zeroes instead of NA).